+++ example/lib/main.dart
@@ -31,8 +31,8 @@ class _MyAppState extends State<MyApp> {
     // Platform messages may fail, so we use a try/catch PlatformException.
     // We also handle the message potentially returning null.
     try {
-      platformVersion =
-          await _flutterViewmodelPlugin.getPlatformVersion() ?? 'Unknown platform version';
+      platformVersion = await _flutterViewmodelPlugin.getPlatformVersion() ??
+          'Unknown platform version';
     } on PlatformException {
       platformVersion = 'Failed to get platform version.';
     }
@@ -60,4 +60,11 @@ class _MyAppState extends State<MyApp> {
       ),
     );
   }
+}
+
+final class TestData {
+  final int id;
+  final String name;
+
+  TestData({required this.id, required this.name});
 }