+++ lib/bases/crud_repository_factory.dart
@@ -1,0 +1,3 @@
+abstract class CrudRepositoryFactory<M> {
+  M generate();
+}